+++ runner/android_test_orchestrator/CHANGELOG.md
@@ -8,6 +8,9 @@
 
 **New Features**
 
+* Introduce Instrumentation Params Proxying which allows the user to proxy
+  instrumentation arguments to the APK under test (eg: --no-hidden-api-checks).
+
 **Breaking Changes**
 
 **API Changes**
@@ -15,3 +18,4 @@
 **Breaking API Changes**
 
 **Known Issues**
+